requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
asgiref=3.8.1=pypi_0
backports-zoneinfo=0.2.1=pypi_0
bzip2=1.0.8=h93a5062_5
ca-certificates=2024.2.2=hf0a4a13_0
django=4.2.11=pypi_0
djangorestframework=3.15.1=pypi_0
libffi=3.4.2=h3422bc3_5
libsqlite=3.45.3=h091b4b1_0
libzlib=1.2.13=h53f4e23_5
ncurses=6.4.20240210=h078ce10_0
openssl=3.3.0=h0d3ecfb_0
pip=24.0=pyhd8ed1ab_0
python=3.8.19=h2469fbe_0_cpython
readline=8.2=h92ec313_1
setuptools=69.5.1=pyhd8ed1ab_0
sqlparse=0.5.0=pypi_0
tk=8.6.13=h5083fa2_1
typing-extensions=4.11.0=pypi_0
wheel=0.43.0=pyhd8ed1ab_1
xz=5.2.6=h57fd34a_0
